Refactor seed response population files
