+++ prisma/seed/responses/helper.ts
@@ -1,7 +1,11 @@
-import { formSelect } from "../../src/forms/forms.service";
-import { prisma } from "./prisma-client";
+import { prisma } from "../prisma-client";
 
-// TODO: move these to a  helper function file
+/*
+generates a random option IDs (in an array) given an option group id
+
+params:
+numberOfChoices - size of the returned array
+ */
 const getRandomOptionId = async (
     optionGroupId: number,
     numberOfChoices: number,
@@ -21,6 +25,10 @@ const getRandomOptionId = async (
     return choicesArray;
 };
 
+/*
+gets all members in the voyage member with the given teamMemberId
+this is used for input type "TeamMemberCheckboxes"
+*/
 const getTeamMembers = async (teamMemberId: number) => {
     const team = await prisma.voyageTeam.findFirst({
         where: {
@@ -45,10 +53,18 @@ const getTeamMembers = async (teamMemberId: number) => {
     return team.voyageTeamMembers.map((m) => m.member.discordId);
 };
 
-const populateQuestionResponses = async (
+/*
+populates responses for a given question
+
+params:
+question - question (which includes id, inputType, etc)
+responseGroupId - this is the group where all the responses for this particular submission are linked to
+teamMemberId - optional, defaulted to 0, only used for teamMember inputType
+*/
+export const populateQuestionResponses = async (
     question: any,
-    teamMemberId: number,
     responseGroupId: number,
+    teamMemberId: number = 0, // this is only used if it's a teamMember type input
 ) => {
     const data: any = {
         questionId: question.id,
@@ -93,7 +109,7 @@ const populateQuestionResponses = async (
                     optionGroupId: true,
                 },
             });
-            // popuate all subquestions
+            // populate all subquestions
             await Promise.all(
                 subQuestions.map((subq) => {
                     // assign subquestion optionGroupId to be same as parent, as it's null for subquestions
@@ -130,6 +146,13 @@ const populateQuestionResponses = async (
             break;
         }
         case "teamMembersCheckbox": {
+            if (teamMemberId === 0) {
+                console.log(question);
+                throw new Error(
+                    `teamMemberId required for input type ${question.inputType.name} (question id:${question.id}).`,
+                );
+            }
+
             const selectedTeamMembers = await getTeamMembers(teamMemberId);
             await prisma.response.create({
                 data: {
@@ -163,72 +186,4 @@ const populateQuestionResponses = async (
             throw new Error("Prisma seed: Unexpected question type");
         }
     }
-};
-
-export const populateCheckinFormResponse = async () => {
-    const teamMemberId = 1;
-    const teamMember = await prisma.voyageTeamMember.findUnique({
-        where: {
-            id: teamMemberId,
-        },
-        select: {
-            id: true,
-            voyageTeam: {
-                select: {
-                    voyage: {
-                        select: {
-                            sprints: true,
-                        },
-                    },
-                },
-            },
-        },
-    });
-
-    // get questions
-    const checkinForm = await prisma.form.findUnique({
-        where: {
-            title: "Sprint Check-in",
-        },
-        select: formSelect,
-    });
-
-    const questions = await prisma.question.findMany({
-        where: {
-            formId: checkinForm.id,
-            parentQuestionId: null,
-        },
-        select: {
-            id: true,
-            order: true,
-            inputType: {
-                select: {
-                    name: true,
-                },
-            },
-            optionGroupId: true,
-            parentQuestionId: true,
-        },
-        orderBy: {
-            order: "asc",
-        },
-    });
-
-    const responseGroup = await prisma.responseGroup.create({
-        data: {},
-    });
-
-    await Promise.all(
-        questions.map((question) => {
-            populateQuestionResponses(question, teamMemberId, responseGroup.id);
-        }),
-    );
-
-    await prisma.formResponseCheckin.create({
-        data: {
-            voyageTeamMemberId: teamMember.id,
-            sprintId: teamMember.voyageTeam.voyage.sprints[0].id,
-            responseGroupId: responseGroup.id,
-        },
-    });
 };
